requirements.txt
astroid==1.6.1
beautifulsoup4==4.6.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.4.1
Django==2.0.2
djangorestframework==3.7.7
google==2.0.1
google-api-python-client==1.6.5
google-auth==1.3.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gunicorn==19.6.0
httplib2==0.10.3
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==4.1.2
oauthlib==2.0.6
# psycopg2==2.6.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pylint==1.8.2
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.2
wrapt==1.10.11
markdown
django-filter